code optimizations
